+++ src/utils/mod.rs
@@ -7,7 +7,7 @@ pub mod time;
 pub mod uri;
 
 pub use base64::{base64_encode, base64_encode_into};
-pub use hasher::{sha256, sha256_bytes};
+pub use hasher::{Sha256, sha256, sha256_bytes};
 pub use mime::guess_mime_type;
 pub use time::{format_iso8601_utc, format_iso8601_utc_into};
 pub use uri::{file_to_data_uri, to_base64_data_uri, to_base64_data_uri_into};